Crate pallet_assets_freezer

Assets Freezer Pallet

A pallet capable of freezing fungibles from pallet-assets. This is an extension of pallet-assets, wrapping fungibles::Inspect. It implements both fungibles::freeze::Inspect and fungibles::freeze::Mutate. The complexity of the operations is O(n). where n is the variant count of RuntimeFreezeReason.

Pallet API

See the pallet module for more information about the interfaces this pallet exposes, including its configuration trait, dispatchables, storage items, events and errors.

Overview

This pallet provides the following functionality:

Pallet hooks allowing [pallet-assets] to know the frozen balance for an account on a given asset (see pallet_assets::FrozenBalance).

An implementation of fungibles::freeze::Inspect and fungibles::freeze::Mutate, allowing other pallets to manage freezes for the pallet-assets assets.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.